added imagesprogram dir with framework for creating text file with data from images

+++ src/pca.h
@@ -3,8 +3,15 @@
 
 #include <Eigen/Dense>
 
-Eigen::MatrixXf getCovarianceMatrix(Eigen::MatrixXf X);
-Eigen::MatrixXf pcaComponents(Eigen::MatrixXf X, int components);
-Eigen::MatrixXf pcaFraction(Eigen::MatrixXf X, double fraction);
+/* returns covariance matrix of variables being columns of X */
+Eigen::MatrixXd getCovarianceMatrix(Eigen::MatrixXd X);
+
+/* Standarize X before using any PCA function! */
+/* Functions below return the feature matrix, to recast the data multiply data_matrix * feature_matrix */
+
+/* returns feature matrix with n_columns = components */
+Eigen::MatrixXd pcaComponents(Eigen::MatrixXd X, int components);
+/* returns feature matrix with enough columns to preserve at least <fraction> of information */
+Eigen::MatrixXd pcaFraction(Eigen::MatrixXd X, double fraction);
 
 #endif
